Improve tevcat code
