+++ docs/tevcat/tevcat.rst
@@ -1,3 +1,5 @@
+.. doctest-skip-all
+
 .. _astroquery.tevcat:
 
 ****************************
